requirements.txt
streamlit
google-generativeai
python-dotenv
langchain
langchain-google-genai
PyPDF2
chromadb
faiss-cpu
pdf2image
